requirements.txt
datasets==2.19.2
gdown==5.2.0
langchain==0.2.4
langchain_community==0.2.4
langchain_core==0.2.6
langchain_huggingface==0.0.3
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.35.0
textsum==0.2.1
chromadb==0.5.0